Allow ChewieController.of
And make the provider private
