+++ example/lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:chewie/chewie.dart';
-import 'package:chewie/src/chewie_controller.dart';
+import 'package:chewie/src/chewie_player.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:video_player/video_player.dart';
